feat: use blacksmith optimization

+++ .bin/scripts/app-build.sh
@@ -35,9 +35,14 @@ get_channel() {
   echo $channel
 }
 
-set +e
-docker buildx create --name mna-lba --driver docker-container --config "$SCRIPT_DIR/buildkitd.toml" 2> /dev/null
-set -e
+if [[ -z "${CI:-}" ]]; then
+  set +e
+  docker buildx create --name mna-lba --driver docker-container --config "$SCRIPT_DIR/buildkitd.toml" 2> /dev/null
+  set -e
+  BUILDER_ARGS="--builder mna-lba"
+else
+  BUILDER_ARGS=""
+fi
 
 if [[ ! -z "${CI:-}" ]]; then
   export DEPS_ID=($(md5sum $ROOT_DIR/yarn.lock))
@@ -47,6 +52,9 @@ fi
 
 export CHANNEL=$(get_channel $VERSION)
 
-docker buildx bake --builder mna-lba --${mode} "$environement"
-docker builder prune --builder mna-lba --keep-storage 20GB --force
-docker buildx stop --builder mna-lba
+docker buildx bake $BUILDER_ARGS --${mode} "$environement"
+
+if [[ -z "${CI:-}" ]]; then
+  docker builder prune --builder mna-lba --keep-storage 20GB --force
+  docker buildx stop --builder mna-lba
+fi
